backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.39
asyncpg==0.31.0
psycopg2-binary==2.9.11
python-dotenv==1.1.0
google-genai==1.68.0
anthropic==0.86.0
openai==2.29.0
requests==2.32.3
pydantic==2.10.3
alembic==1.18.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
